FARMASATHI/backend/requirements.txt
# Core FastAPI Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Environment & Configuration
python-dotenv==1.0.0

# File Handling
aiofiles==23.2.1
Pillow>=10.3.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Machine Learning (Optional)
scikit-learn==1.3.2
# torch==2.1.1
# torchvision==0.16.1

# HTTP Requests
requests==2.31.0
httpx==0.25.2

# WebSocket
websockets==12.0

# Utilities
typing-extensions==4.9.0
email-validator==2.1.0

# AI & Text-to-Speech
openai==1.6.1
google-generativeai==0.3.2
gTTS==2.5.0
pydub==0.25.1
